pyproject.toml
[project]
name = "plotdigitizer"
version = "0.1.0"
description = "Convert plot images to CSV data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pytesseract>=0.3.10",
    "opencv-python-headless>=4.8",
]

[project.scripts]
plot_digitizer = "plot_digitizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["plot_digitizer"]
